Stop formatting based on package name length
Since we have extended approved line length to 120 it is unlikely that imports should be reformatted. For example with that one conditional formatting the package name could be 38 chars long before alternative formatting should occur. This should be a rare case and it's okay to let a user to fix that.